Hook up express api
