+++ api/package.json
@@ -8,7 +8,7 @@
     "start": "nodemon -r dotenv/config -r esm index.js"
   },
   "dependencies": {
-    "apollo-server": "^2.1.0",
+    "apollo-server-express": "^2.1.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.3",
     "cors": "^2.8.4",
