+++ api/strategies/basic.js
@@ -1,10 +1,10 @@
 import bcrypt from 'bcryptjs';
+import db from '../models';
 import {BasicStrategy} from 'passport-http';
-import {User} from '../models';
 
 export default new BasicStrategy(async (email, password, done) => {
   try {
-    const user = await User.findOne({where: {email}});
+    const user = await db.user.findOne({where: {email}});
     if (user) {
       const valid = await bcrypt.compare(password, user.password);
       if (valid) {
