+++ api/strategies/jwt.js
@@ -1,5 +1,5 @@
+import db from '../models';
 import {ExtractJwt, Strategy} from 'passport-jwt';
-import {User} from '../models';
 
 export const jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
 export default new Strategy(
@@ -9,7 +9,7 @@ export default new Strategy(
   },
   async (payload, done) => {
     try {
-      const user = await User.findById(payload.sub);
+      const user = await db.user.findById(payload.sub);
       if (!user) {
         return done(null, false);
       }
