+++ client/.eslintrc
@@ -7,6 +7,7 @@
     "es6": true
   },
   "globals": {
-    "TITLE": true
+    "TITLE": true,
+    "API_URL": true
   }
 }
