+++ client/src/api.js
@@ -1,10 +1,7 @@
 import Frisbee from 'frisbee';
 
 export default new Frisbee({
-  baseURI:
-    process.env.NODE_ENV === 'production'
-      ? 'https://api.batbstats.trevorblades.com'
-      : 'http://localhost:3000',
+  baseURI: API_URL,
   headers: {
     Accept: 'application/json',
     'Content-Type': 'application/json'
