+++ client/webpack.config.js
@@ -30,7 +30,12 @@ module.exports = {
       template: 'index.html'
     }),
     new webpack.DefinePlugin({
-      TITLE: JSON.stringify(title)
+      TITLE: JSON.stringify(title),
+      API_URL: JSON.stringify(
+        process.env.NODE_ENV === 'production'
+          ? 'https://api.batbstats.trevorblades.com'
+          : 'http://localhost:3000'
+      )
     })
   ]
 };
